news_bot/requirements.txt
python-telegram-bot==13.15
requests==2.31
beautifulsoup4==4.12.2